# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroHub API
  description: 'API operacional de AeroHub: aprovisionamiento de tenants, seguimiento
    de vuelos y distribucion de FIDS en tiempo real (RF-T02, OpenAPI 3.1).'
  contact:
    name: AeroHub Platform Team
  version: 0.1.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /rampa/turnarounds:
    get:
      tags:
      - rampa
      summary: Listar Turnarounds Endpoint
      operationId: listar_turnarounds_endpoint_rampa_turnarounds_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TurnaroundResponse'
                type: array
                title: Response Listar Turnarounds Endpoint Rampa Turnarounds Get
    post:
      tags:
      - rampa
      summary: Crear Turnaround Endpoint
      description: 'Empareja un vuelo de llegada y uno de salida de la misma aeronave

        en un turnaround (RF-O16, SDD-DATA-001 §8.3).'
      operationId: crear_turnaround_endpoint_rampa_turnarounds_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearTurnaroundRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrearTurnaroundResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rampa/turnarounds/{turnaround_id}/tareas:
    post:
      tags:
      - rampa
      summary: Iniciar Tarea Endpoint
      description: El agente de rampa marca el inicio de una tarea (CU-O16, paso 2).
      operationId: iniciar_tarea_endpoint_rampa_turnarounds__turnaround_id__tareas_post
      parameters:
      - name: turnaround_id
        in: path
        required: true
        schema:
          type: integer
          title: Turnaround Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IniciarTareaRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IniciarTareaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - rampa
      summary: Listar Tareas Endpoint
      description: 'Minimo privilegio (Plan §8.5): role_ramp_agent solo ve sus propias

        tareas -- filtrado en la capa de aplicacion/infraestructura.'
      operationId: listar_tareas_endpoint_rampa_turnarounds__turnaround_id__tareas_get
      parameters:
      - name: turnaround_id
        in: path
        required: true
        schema:
          type: integer
          title: Turnaround Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TareaResponse'
                title: Response Listar Tareas Endpoint Rampa Turnarounds  Turnaround
                  Id  Tareas Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rampa/tareas/{tarea_id}/finalizar:
    post:
      tags:
      - rampa
      summary: Finalizar Tarea Endpoint
      description: 'El agente de rampa marca el fin de una tarea; genera una incidencia

        de rampa automaticamente si supera el estandar (CU-O16 pasos 3-4,

        RF-O16). Minimo privilegio: 404 (no 403) si la tarea es de otro agente

        y el rol activo es role_ramp_agent -- ver aerohub_ramp.application.

        finalizar_tarea (PN-01).'
      operationId: finalizar_tarea_endpoint_rampa_tareas__tarea_id__finalizar_post
      parameters:
      - name: tarea_id
        in: path
        required: true
        schema:
          type: integer
          title: Tarea Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizarTareaRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizarTareaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /rampa/informes/simple:
    get:
      tags:
      - rampa
      summary: Informe Turnarounds Simple Endpoint
      operationId: informe_turnarounds_simple_endpoint_rampa_informes_simple_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: estado
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Estado
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /rampa/informes/compuesto:
    get:
      tags:
      - rampa
      summary: Informe Turnarounds Compuesto Endpoint
      operationId: informe_turnarounds_compuesto_endpoint_rampa_informes_compuesto_get
      parameters:
      - name: periodo_inicio
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Inicio
      - name: periodo_fin
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Periodo Fin
      - name: formato
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Formato
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CrearTurnaroundRequest:
      properties:
        vuelo_llegada_id:
          type: integer
          title: Vuelo Llegada Id
        vuelo_salida_id:
          type: integer
          title: Vuelo Salida Id
        inicio_previsto:
          type: string
          format: date-time
          title: Inicio Previsto
        fin_previsto:
          type: string
          format: date-time
          title: Fin Previsto
      type: object
      required:
      - vuelo_llegada_id
      - vuelo_salida_id
      - inicio_previsto
      - fin_previsto
      title: CrearTurnaroundRequest
    CrearTurnaroundResponse:
      properties:
        turnaround_id:
          type: string
          title: Turnaround Id
      type: object
      required:
      - turnaround_id
      title: CrearTurnaroundResponse
...
    FinalizarTareaRequest:
      properties:
        fin_real:
          type: string
          format: date-time
          title: Fin Real
      type: object
      required:
      - fin_real
      title: FinalizarTareaRequest
    FinalizarTareaResponse:
      properties:
        tarea_id:
          type: string
          title: Tarea Id
        duracion_minutos:
          type: number
          title: Duracion Minutos
        incidencia_generada:
          type: boolean
          title: Incidencia Generada
      type: object
      required:
      - tarea_id
      - duracion_minutos
      - incidencia_generada
      title: FinalizarTareaResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IniciarTareaRequest:
      properties:
        tipo_tarea_id:
          type: integer
          title: Tipo Tarea Id
      type: object
      required:
      - tipo_tarea_id
      title: IniciarTareaRequest
    IniciarTareaResponse:
      properties:
        tarea_id:
          type: string
          title: Tarea Id
      type: object
      required:
      - tarea_id
      title: IniciarTareaResponse
...
    TareaResponse:
      properties:
        id:
          type: string
          title: Id
        turnaround_id:
          type: string
          title: Turnaround Id
        tipo_tarea_id:
          type: string
          title: Tipo Tarea Id
        agente_usuario_id:
          type: string
          title: Agente Usuario Id
        inicio_real:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Inicio Real
        fin_real:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Fin Real
        estado:
          type: string
          title: Estado
      type: object
      required:
      - id
      - turnaround_id
      - tipo_tarea_id
      - agente_usuario_id
      - inicio_real
      - fin_real
      - estado
      title: TareaResponse
...
    TurnaroundResponse:
      properties:
        id:
          type: string
          title: Id
        vuelo_llegada_id:
          type: string
          title: Vuelo Llegada Id
        numero_vuelo_llegada:
          type: string
          title: Numero Vuelo Llegada
        vuelo_salida_id:
          type: string
          title: Vuelo Salida Id
        numero_vuelo_salida:
          type: string
          title: Numero Vuelo Salida
        aeronave_id:
          type: string
          title: Aeronave Id
        inicio_previsto:
          type: string
          format: date-time
          title: Inicio Previsto
        fin_previsto:
          type: string
          format: date-time
          title: Fin Previsto
        estado:
          type: string
          title: Estado
      type: object
      required:
      - id
      - vuelo_llegada_id
      - numero_vuelo_llegada
      - vuelo_salida_id
      - numero_vuelo_salida
      - aeronave_id
      - inicio_previsto
      - fin_previsto
      - estado
      title: TurnaroundResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...